third_party/WebKit/LayoutTests/fast/css/border-radius-property-value-expected.txt

CONSOLE WARNING: 'getMatchedCSSRules()' is deprecated. For more help, check https://code.google.com/p/chromium/issues/detail?id=437569#c2
Test for Bug: 80736 - CSSStyleDeclaration.getPropertyValue() for 'border-radius' returns null when value is set

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS style.getPropertyValue('border-radius') is '5px'
PASS style.getPropertyValue('border-top-left-radius') is '5px'
PASS style.getPropertyValue('border-top-right-radius') is '5px'
PASS style.getPropertyValue('border-bottom-left-radius') is '5px'
PASS style.getPropertyValue('border-bottom-right-radius') is '5px'
PASS successfullyParsed is true

TEST COMPLETE
